+++ packages/perseus-core/src/index.ts
@@ -16,4 +16,4 @@ export {PerseusError} from "./error/perseus-error";
 
 export * from "./data-schema";
 
-export {pluck, mapObject} from "./utils/objective_";
+export {pluck, mapObject, clone} from "./utils/object-utils";
