+++ packages/perseus/src/__tests__/renderer.test.tsx
@@ -1,10 +1,9 @@
 import {describe, beforeAll, beforeEach, it} from "@jest/globals";
-import {Errors} from "@khanacademy/perseus-core";
+import {Errors, clone} from "@khanacademy/perseus-core";
 import {act, screen, waitFor, within} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
-import {clone} from "../../../../testing/object-utils";
 import {testDependencies} from "../../../../testing/test-dependencies";
 import {
     dropdownWidget,
