+++ packages/perseus/src/widgets/interactive-graphs/interactive-graph.test.tsx
@@ -1,5 +1,5 @@
 import {describe, beforeEach, it} from "@jest/globals";
-import {lockedFigureColors} from "@khanacademy/perseus-core";
+import {lockedFigureColors, clone} from "@khanacademy/perseus-core";
 import {color as wbColor} from "@khanacademy/wonder-blocks-tokens";
 import {act, waitFor} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
@@ -7,7 +7,6 @@ import {Plot} from "mafs";
 import * as React from "react";
 import invariant from "tiny-invariant";
 
-import {clone} from "../../../../../testing/object-utils";
 import {testDependencies} from "../../../../../testing/test-dependencies";
 import {waitForInitialGraphieRender} from "../../../../../testing/wait";
 import {getDefaultFigureForType} from "../../../../perseus-editor/src/widgets/interactive-graph-editor/locked-figures/util";
